+++ great_international/apps.py
@@ -36,6 +36,8 @@ class GreatInternationalConfig(AppConfig):
         cache.AboutDitServicesPageSubscriber.subscribe()
         cache.AboutUkLandingPageSubscriber.subscribe()
         cache.AboutUkWhyChooseTheUkPageSubscriber.subscribe()
+        cache.CapitalInvestContactFormPageSubscriber.subscribe()
+        cache.CapitalInvestContactFormSuccessPageSubscriber.subscribe()
         # tags inheritance signals
         post_save.connect(
             receiver=signals.inherit_tags_from_parent,