+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## Pre-release
 
 ### Implemented enhancements
+- No ticket - Make teaser field optional in international articles
 - No ticket - Add featured industries to Invest home page
 - CI-321 - About UK landing page
+- CI-276 - Added `CapitalInvestContactFormPage` and `CapitalInvestContactFormSuccessPage`
+- CI-429 - Tree based breadcrumbs can now use `breadcrumbs_label` if available 
 
 ### Fixed Bugs
 - CI-426 - Added pdf document upload to why choose uk page for ebook section